refactor: change file name to password-hashing
